python_backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
langchain==0.3.13
langchain-groq==0.2.1
langchain-community==0.3.13
tavily-python==0.5.0
pydantic==2.10.3
pydantic-settings==2.6.1
pyautogen==0.10.0
